% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helminthiasis.R
\name{paradrug_helminthiasis}
\alias{paradrug_helminthiasis}
\title{Analysis of Helminthiasis}
\usage{
paradrug_helminthiasis(
  object,
  Rbas = "BL_KK2_AL_EPG",
  Rfol = "FU_KK2_AL_EPG",
  Tbas = "BL_KK2_TT_EPG",
  Tfol = "FU_KK2_TT_EPG",
  Hbas = "BL_KK2_HW_EPG",
  Hfol = "FU_KK2_HW_EPG",
  ...
)
}
\arguments{
\item{object}{an object of class paradrug_rawdata as returned by \code{\link{read_paradrug_xls}}}

\item{Rbas}{column in name in object$data for Rbas/Rfol: Ascaris lumbricoides, in eggs per gram of stool - BASELINE/FOLLOW-UP}

\item{Rfol}{column in name in object$data for Rbas/Rfol: Ascaris lumbricoides, in eggs per gram of stool - BASELINE/FOLLOW-UP}

\item{Tbas}{column in name in object$data for Tbas/Tfol: Trichuris trichiura, in eggs per gram of stool - BASELINE/FOLLOW-UP}

\item{Tfol}{column in name in object$data for Tbas/Tfol: Trichuris trichiura, in eggs per gram of stool - BASELINE/FOLLOW-UP}

\item{Hbas}{column in name in object$data for Hbas/Hfol: Hookworms, in eggs per gram of stool - BASELINE/FOLLOW-UP}

\item{Hfol}{column in name in object$data for Hbas/Hfol: Hookworms, in eggs per gram of stool - BASELINE/FOLLOW-UP}

\item{...}{not used yet}
}
\value{
TODO
}
\description{
Analysis of Helminthiasis
}
\examples{
path <- system.file(package = "ParaDrug", "extdata", "data", "mydata.xlsx")
x <- read_paradrug_xls(path)
p <- paradrug_helminthiasis(x)
}
