% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helminthiasis.R
\name{paradrug_helminthiasis_conclusion}
\alias{paradrug_helminthiasis_conclusion}
\title{Conclusion of Helminthiasis}
\usage{
paradrug_helminthiasis_conclusion(
  object,
  Rbas = "BL_KK2_AL_EPG",
  Rfol = "FU_KK2_AL_EPG",
  Tbas = "BL_KK2_TT_EPG",
  Tfol = "FU_KK2_TT_EPG",
  Hbas = "BL_KK2_HW_EPG",
  Hfol = "FU_KK2_HW_EPG",
  drug = c("Albendazole (1x 400 mg)", "Mebendazole (1x 500 mg)", "Other"),
  type = c("latex", "markdown"),
  ...
)
}
\arguments{
\item{object}{an object of class paradrug_rawdata as returned by \code{\link{read_paradrug_xls}}}

\item{Rbas}{column in name in object$data for Rbas/Rfol: Ascaris lumbricoides, in eggs per gram of stool - BASELINE/FOLLOW-UP}

\item{Rfol}{column in name in object$data for Rbas/Rfol: Ascaris lumbricoides, in eggs per gram of stool - BASELINE/FOLLOW-UP}

\item{Tbas}{column in name in object$data for Tbas/Tfol: Trichuris trichiura, in eggs per gram of stool - BASELINE/FOLLOW-UP}

\item{Tfol}{column in name in object$data for Tbas/Tfol: Trichuris trichiura, in eggs per gram of stool - BASELINE/FOLLOW-UP}

\item{Hbas}{column in name in object$data for Hbas/Hfol: Hookworms, in eggs per gram of stool - BASELINE/FOLLOW-UP}

\item{Hfol}{column in name in object$data for Hbas/Hfol: Hookworms, in eggs per gram of stool - BASELINE/FOLLOW-UP}

\item{drug}{either "Albendazole (1x 400 mg)", "Mebendazole (1x 500 mg)" or "Other"}

\item{type}{either 'latex' or 'markdown' to return a latex representation of the analysis or a markdown version. Defaults to 'latex'.}

\item{...}{not used yet}
}
\value{
TODO
}
\description{
Conclusion of Helminthiasis
}
\examples{
path <- system.file(package = "ParaDrug", "extdata", "data", "mydata.xlsx")
x <- read_paradrug_xls(path)
p <- paradrug_helminthiasis_conclusion(x, drug = "Albendazole (1x 400 mg)")
p <- paradrug_helminthiasis_conclusion(x, drug = "Mebendazole (1x 500 mg)")
p <- paradrug_helminthiasis_conclusion(x, drug = "Other")
}
