% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{paradrug_report}
\alias{paradrug_report}
\title{Create a Paradrug report}
\usage{
paradrug_report(x, params = list(), ...)
}
\arguments{
\item{x}{an object as returned by \code{\link{read_paradrug_xls}}}

\item{params}{a list of options with elements}

\item{...}{not used yet}
}
\value{
a list with elements
\itemize{
\item{report: the source file of the report}
\item{out: the pdf of the report}
\item{params: the list of params for the report}
}
}
\description{
Create a Paradrug report
}
\examples{
path   <- system.file(package = "ParaDrug", "extdata", "data", "mydata.xlsx")
x      <- read_paradrug_xls(path)
library(xlsx)
params <- list(
    Country = "a not further specified country", 
    Name = "Unknown", 
    Region = "district/province not further specified",
    file1 = list(datapath = system.file(package = "ParaDrug", "extdata", "data", "mydata.xlsx")),
    NTD = 1, Sdrug = 1, STHdrug = 1, 
    Shbas = "BL_KK2_AL_EPG", Shfol = "FU_KK2_AL_EPG", 
    Smbas = "BL_KK2_TT_EPG", Smfol = "FU_KK2_TT_EPG", 
    Sjbas = "BL_KK2_HW_EPG", Sjfol = "FU_KK2_HW_EPG", 
    Rbas = "Not recorded", Rfol = "Not recorded", 
    Tbas = "Not recorded", Tfol = "Not recorded", 
    Hbas = "Not recorded", Hfol = "Not recorded",
    followup = "Not recorded")
report <- paradrug_report(x, params = params)
}
