% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schistosomiasis.R
\name{paradrug_schistosomiasis_conclusion}
\alias{paradrug_schistosomiasis_conclusion}
\title{Conclusion of Schistosomiasis}
\usage{
paradrug_schistosomiasis_conclusion(
  object,
  Shbas = "BL_KK2_AL_EPG",
  Shfol = "FU_KK2_AL_EPG",
  Smbas = "BL_KK2_TT_EPG",
  Smfol = "FU_KK2_TT_EPG",
  Sjbas = "BL_KK2_HW_EPG",
  Sjfol = "FU_KK2_HW_EPG",
  drug = c("Praziquantel (1x 40 mg/kg)", "Other"),
  type = c("latex", "markdown"),
  ...
)
}
\arguments{
\item{object}{an object of class paradrug_rawdata as returned by \code{\link{read_paradrug_xls}}}

\item{Shbas}{column in name in object$data for Shbas/Shfol: S. haematobium, in eggs per 10 ml of urine - BASELINE/FOLLOW-UP}

\item{Shfol}{column in name in object$data for Shbas/Shfol: S. haematobium, in eggs per 10 ml of urine - BASELINE/FOLLOW-UP}

\item{Smbas}{column in name in object$data for Smbas/Smfol: S. mansoni, in eggs per gram of stool - BASELINE/FOLLOW-UP}

\item{Smfol}{column in name in object$data for Smbas/Smfol: S. mansoni, in eggs per gram of stool - BASELINE/FOLLOW-UP}

\item{Sjbas}{column in name in object$data for Shbas/Shfol: Sjbas/Sjfol: S. japonicum, in eggs per gram of stool - BASELINE/FOLLOW-UP}

\item{Sjfol}{column in name in object$data for Shbas/Shfol: Sjbas/Sjfol: S. japonicum, in eggs per gram of stool - BASELINE/FOLLOW-UP}

\item{drug}{either "Praziquantel (1x 40 mg/kg)" or "Other"}

\item{type}{either 'latex' or 'markdown' to return a latex representation of the analysis or a markdown version. Defaults to 'latex'.}

\item{...}{not used yet}
}
\value{
a character string with the general conclusion of the trial
}
\description{
Conclusion of Schistosomiasis
}
\examples{
path <- system.file(package = "ParaDrug", "extdata", "data", "mydata.xlsx")
x <- read_paradrug_xls(path)
p <- paradrug_schistosomiasis_conclusion(x, drug = "Praziquantel (1x 40 mg/kg)")
p <- paradrug_schistosomiasis_conclusion(x, drug = "Other")
p <- paradrug_schistosomiasis_conclusion(x, drug = "Praziquantel (1x 40 mg/kg)",
                                         type = "markdown")
p <- paradrug_schistosomiasis_conclusion(x, drug = "Other",
                                         type = "markdown")
}
