% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paradrug-shiny.R
\name{paradrugShiny}
\alias{paradrugShiny}
\alias{paradrugUI}
\alias{paradrugServer}
\title{Paradrug Shiny app component}
\usage{
paradrugUI(id, ...)

paradrugServer(id, ...)
}
\arguments{
\item{id}{character string with the id of the module}

\item{...}{not used}
}
\value{
\code{paradrugUI} returns the the ui elements of the Paradrug app.\cr
\code{paradrugServer} returns the the server components of the Paradrug app.
}
\description{
Main UI part of the Paradrug app
}
\examples{
if(interactive()){
library(ParaDrug)

ui <- fluidPage(
  paradrugUI(id = "activity")
)
server <- function(input, output, session) {
  paradrugServer(id = "activity")
}
shinyApp(ui, server)
}
}
