% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-input.R
\name{read_paradrug_xls}
\alias{read_paradrug_xls}
\title{Read xlsx data with paradrug information}
\usage{
read_paradrug_xls(x, sheet = 1)
}
\arguments{
\item{x}{character string with the path to the file}

\item{sheet}{either the character string with the name of the sheet or the index number of the sheet. Defaults to the first sheet.}
}
\value{
a list with elements 
\itemize{
\item{data: data.frame with at least columns "Site", "SubjectID", "SchoolID", "Age", "Sex2", "BL_KK2_AL_EPG", "FU_KK2_AL_EPG", "BL_KK2_TT_EPG", "FU_KK2_TT_EPG", "BL_KK2_HW_EPG", "FU_KK2_HW_EPG"}
\item{n: the number of rows in the dataset} 
}
}
\description{
Read xlsx data with paradrug information
}
\examples{
path <- system.file(package = "ParaDrug", "extdata", "data", "mydata.xlsx")
x <- read_paradrug_xls(path)
}
