% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helminthiasis.R
\name{plot_paradrug_helminthiasis_eggcount_reduction}
\alias{plot_paradrug_helminthiasis_eggcount_reduction}
\title{Plot of eggcount reduction of Helminthiasis}
\usage{
plot_paradrug_helminthiasis_eggcount_reduction(
  object,
  Rbas = "BL_KK2_AL_EPG",
  Rfol = "FU_KK2_AL_EPG",
  Tbas = "BL_KK2_TT_EPG",
  Tfol = "FU_KK2_TT_EPG",
  Hbas = "BL_KK2_HW_EPG",
  Hfol = "FU_KK2_HW_EPG",
  drug = c("Albendazole (1x 400 mg)", "Mebendazole (1x 500 mg)", "Other"),
  ...
)
}
\arguments{
\item{object}{an object of class paradrug_rawdata as returned by \code{\link{read_paradrug_xls}}}

\item{Rbas}{column in name in object$data for Rbas/Rfol: Ascaris lumbricoides, in eggs per gram of stool - BASELINE/FOLLOW-UP}

\item{Rfol}{column in name in object$data for Rbas/Rfol: Ascaris lumbricoides, in eggs per gram of stool - BASELINE/FOLLOW-UP}

\item{Tbas}{column in name in object$data for Tbas/Tfol: Trichuris trichiura, in eggs per gram of stool - BASELINE/FOLLOW-UP}

\item{Tfol}{column in name in object$data for Tbas/Tfol: Trichuris trichiura, in eggs per gram of stool - BASELINE/FOLLOW-UP}

\item{Hbas}{column in name in object$data for Hbas/Hfol: Hookworms, in eggs per gram of stool - BASELINE/FOLLOW-UP}

\item{Hfol}{column in name in object$data for Hbas/Hfol: Hookworms, in eggs per gram of stool - BASELINE/FOLLOW-UP}

\item{drug}{either "Albendazole (1x 400 mg)", "Mebendazole (1x 500 mg)" or "Other"}

\item{...}{not used yet}
}
\value{
a plot of the egg count reduction in the trial is generated
}
\description{
Plot of eggcount reduction of Helminthiasis
}
\examples{
path <- system.file(package = "ParaDrug", "extdata", "data", "mydata.xlsx")
x <- read_paradrug_xls(path)
p <- plot_paradrug_helminthiasis_eggcount_reduction(x, drug = "Albendazole (1x 400 mg)")
p <- plot_paradrug_helminthiasis_eggcount_reduction(x, drug = "Mebendazole (1x 500 mg)")
p <- plot_paradrug_helminthiasis_eggcount_reduction(x, drug = "Other")
p <- plot_paradrug_helminthiasis_eggcount_reduction(x, drug = "Albendazole (1x 400 mg)", 
  Rbas = "BL_KK2_AL_EPG", Rfol = "FU_KK2_AL_EPG", 
  Tbas = "BL_KK2_TT_EPG", Tfol = "FU_KK2_TT_EPG",  
  Hbas = "Not recorded", Hfol = "Not recorded")
p <- plot_paradrug_helminthiasis_eggcount_reduction(x, drug = "Albendazole (1x 400 mg)", 
  Rbas = "BL_KK2_AL_EPG", Rfol = "FU_KK2_AL_EPG", 
  Tbas = "Not recorded", Tfol = "Not recorded",  
  Hbas = "Not recorded", Hfol = "Not recorded")
}
